split dashboard into component files
